interface is now truly just the public interface. users should only import one or the other
